トースターの修正

+++ src/app/lib/actions.ts
@@ -4,7 +4,7 @@ import { parseWithZod } from "@conform-to/zod";
 import { PrismaClient } from "@prisma/client";
 import { revalidatePath } from "next/cache";
 
-import { postSchema } from "../schema";
+import { deletePostSchema, postSchema } from "../schema";
 
 const prisma = new PrismaClient();
 
@@ -71,9 +71,17 @@ export async function editPost(prevState: unknown, formData: FormData) {
   // redirect("/");
 }
 
+// ToDo deleteだけ独自に返り値を返すようにしているが、見直したい（formに入力欄がないため）
 // ToDo server actionなのでasyncが必要だが、lintエラーが起きるので下記で一旦無効化している
 // eslint-disable-next-line
 export async function deletePost(ids: number[]) {
+  const submission = deletePostSchema.safeParse({
+    ids,
+  });
+
+  if (!submission.success) {
+    return { status: "error", ids: ids };
+  }
   await prisma.post.deleteMany({
     where: {
       id: {
@@ -83,4 +91,5 @@ export async function deletePost(ids: number[]) {
   });
 
   revalidatePath("/");
+  return { status: "success", ids: ids };
 }
